+++ lua/codecompanion/config.lua
@@ -147,8 +147,8 @@ local defaults = {
           callback = "strategies.chat.agents.tools.next_edit_suggestion",
           description = "Suggest and jump to the next position to edit",
         },
-        ["symbol_context"] = {
-          callback = "strategies.chat.agents.tools.symbol_context",
+        ["list_code_usages"] = {
+          callback = "strategies.chat.agents.tools.list_code_usages",
           description = "Find code symbol context",
         },
         opts = {
@@ -436,6 +436,12 @@ local defaults = {
           index = 19,
           callback = "keymaps.goto_file_under_cursor",
           description = "Open the file under cursor in a new tab.",
+        },
+        copilot_stats = {
+          modes = { n = "gS" },
+          index = 20,
+          callback = "keymaps.copilot_stats",
+          description = "Show Copilot usage statistics",
         },
       },
       opts = {
@@ -1004,8 +1010,8 @@ You must create or modify a workspace file through a series of prompts over mult
     },
     chat = {
       icons = {
-        pinned_buffer = " ",
-        watched_buffer = "󰂥 ",
+        buffer_pin = " ",
+        buffer_watch = "󰂥 ",
       },
       debug_window = {
         ---@return number|fun(): number